add switching drawing and result by user drew posts
