+++ src/features/Drawing/Result.module.css
@@ -1,0 +1,21 @@
+.vis {
+  width: 95vw;
+  height: 64vh;
+  margin: 12px auto;
+}
+
+.download {
+  box-sizing: border-box;
+  display: block;
+  width: 280px;
+  height: 48px;
+  padding: 14px 0;
+  margin: 24px auto;
+  font-size: 1.25rem;
+  font-weight: bold;
+  color: #fff;
+  text-align: center;
+  text-decoration: none;
+  background: var(--theme-color-main);
+  border-radius: 8px;
+}
